feat(dfa): #43 add dfa minimization algorithm
- Implement the `minimize` function for the DFA
class. This funciton is used to reduce the number
of redundant states in the DFA.
- Add test cases for the `minimize` function.
- Fix a bug in the `accepts` function of the DFA
class where the empty string was being accepted
by the DFA even when it was not in a final state.
- Modify some existing functions to make them
private and add type hints to the functions.
- Modify the `minimize` function in the NFA class
to use the `minimize` function of the DFA class
when converting the NFA to a DFA.